requirements.txt
speechrecognition
pyttsx3
google-genai